Int format problem solved by switching to uint64_t.

+++ src/pmdas/linux_zfs/zfs_dmu_tx.h
@@ -1,16 +1,16 @@
 typedef struct zfs_dmu_tx {
-    unsigned int assigned;
-    unsigned int delay;
-    unsigned int error;
-    unsigned int suspended;
-    unsigned int group;
-    unsigned int memory_reserve;
-    unsigned int memory_reclaim;
-    unsigned int dirty_throttle;
-    unsigned int dirty_delay;
-    unsigned int dirty_over_max;
-    unsigned int dirty_frees_delay;
-    unsigned int quota;
+    uint64_t assigned;
+    uint64_t delay;
+    uint64_t error;
+    uint64_t suspended;
+    uint64_t group;
+    uint64_t memory_reserve;
+    uint64_t memory_reclaim;
+    uint64_t dirty_throttle;
+    uint64_t dirty_delay;
+    uint64_t dirty_over_max;
+    uint64_t dirty_frees_delay;
+    uint64_t quota;
 } zfs_dmu_tx_t;
 
 void zfs_dmu_tx_refresh(zfs_dmu_tx_t *dmu_tx, regex_t *rgx_row);
