+++ src/pmdas/linux_zfs/zfs_zfetchstats.h
@@ -1,7 +1,7 @@
 typedef struct zfs_zfetchstats {
-    unsigned int hits;
-    unsigned int misses;
-    unsigned int max_streams;
+    uint64_t hits;
+    uint64_t misses;
+    uint64_t max_streams;
 } zfs_zfetchstats_t;
 
 void zfs_zfetchstats_refresh(zfs_zfetchstats_t *zfetchstats, regex_t *rgx_row);
